Skip aliases of already verified domains.
When issuing a two-domain certificate using a different alias for each
domain, if the first domain is already verified, verification for the
second domain would be attempted (unsuccessfully) using the alias of the
first domain.

Increment the alias index when skipping verified domains so that the
correct alias will be used for subsequent domains.